encryption and decryption of data
Video Steganography is a technique to hide any kind of files in any extension into a carrying video file.  In this project, an application developed to embed any kind of data(File) into another file, which is called carrier file.  The carrier file must be a video file.  Even though there are numerous technological advances in video steganography, the software we are creating is a newly developed software version of our ideas.